pyproject.toml
[project]
name = "smol-notebooklm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "openai>=1.90.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.0",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.0",
    "together>=1.5.17",
]
